// frontend/src/app/shared/generated/models/update-analysis-template-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateAnalysisTemplateRequest {
  calculationDefinitions?: (string | null);
  maxTolerance?: (number | string | null);
  minTolerance?: (number | string | null);
  name: string;
  rowVersion: string;
  testConfiguration?: (string | null);
  validationRules?: (string | null);
}